[8493] Tracker choices from  selected projects in filtering
The first release of Ethan (the super cool search function for
Belvedere's Redmine site) had a bug where if you did a search,
then changed the "Projects" dropdown to be something that had
more trackers than what you just searched for, it would not add
these tracker types to the search.

This fixes the tracker checkboxes so that they will display
according to what projects you have selected, i.e. if you are
on a project with "Bug" and "Feature" then you switch the
projects you are searching to "All projects" and there are
other projects that have a "Task" tracker, you will see "Bug",
"Feature", "Task".

Change-Id: I248fa1a5bb26fe9968f0932fcc5f1ff3e9f89d79